Atomic ID allocation
A Wayland compositor implemented using libwayland expects sequential
IDs. That is, when calling a constructor, we must use ID n+1, not n+2,
or the connection will be terminated.

Because we allocated IDs in NewProxy, there was a race where two
objects may get sequential IDs, but the second constructor gets called
before the first constructor, which looks like a non-sequential ID to
the server.

Instead, move ID allocation into SendRequest, which already holds a
mutex.

This does simplify the API a bit, because we can also move the
remainder of NewProxy into SendRequest, which primarily helps user
code working with wl_registry, which no longer needs to call NewProxy
before calling wl_registry_bind.