cmd/wayland-scanner: generate type-safe wrappers for adding globals
Provide type-safe wrappers for wlserver.Display.AddGlobal that apply
stricter typing to the bind callback.

The downside of this approach is that we generate these functions for
all types, even those that shouldn't be used as globals. The protocol
specifications don't tell us in a machine-readable way whether an
interface gets used as a global or not.